refactoring the base_node
