--- modules/common/base_gui_node.py
@@ -1,8 +1,0 @@
-from modules.executors.gui_node_builder import GuiNodeBuilder
-
-class BaseGuiNode(GuiNodeBuilder):
-
-    def buildNode(self):
-        builder = self._initBuilder()
-        return builder
-
